Moves attachment_helper to rails/attachment_helper
