+++ lib/refile/rails/attachment_helper.rb
@@ -55,5 +55,3 @@ module Refile
     end
   end
 end
-
-ActionView::Base.send(:include, Refile::AttachmentHelper)
